chore(EST-3003-BE): rename modules for brokers and storages
