+++ src/shared/communication/storages/__init__.py
@@ -1,8 +1,8 @@
 """Storage implementations and the AbstractStorage contract."""
 
 from .abstract import AbstractStorage
-from .redis_ import RedisStorage
-from .minio_ import MinioStorage
+from .redis_storage import RedisStorage
+from .minio_storage import MinioStorage
 
 __all__ = [
     "AbstractStorage",
