+++ src/shared/communication/storages/minio_storage.py
